+++ src/main/java/com/codeactuator/rocket/dao/ProjectRepository.java
@@ -4,7 +4,7 @@ package com.codeactuator.rocket.dao;
 import com.codeactuator.rocket.domain.Project;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProjectRepository extends CrudRepository<Project, Integer>, ProjectCustomRepository {
+public interface ProjectRepository extends CrudRepository<Project, Long>, ProjectCustomRepository {
 
 
 }
